+++ src/main/kotlin/io/openfuture/chain/repository/Repositories.kt
@@ -7,6 +7,7 @@ import io.openfuture.chain.entity.Block
 import io.openfuture.chain.entity.Wallet
 import io.openfuture.chain.entity.Delegate
 import io.openfuture.chain.entity.transaction.BaseTransaction
+import io.openfuture.chain.entity.transaction.CoinBaseTransaction
 import io.openfuture.chain.entity.transaction.TransferTransaction
 import io.openfuture.chain.entity.transaction.VoteTransaction
 import org.springframework.data.domain.Pageable
@@ -18,7 +19,7 @@ import org.springframework.stereotype.Repository
 interface BaseRepository<T> : JpaRepository<T, Int>
 
 @Repository
-interface BlockRepository<T: Block> : BaseRepository<T> {
+interface BlockRepository<T : Block> : BaseRepository<T> {
 
     fun findByHash(hash: String): T?
 
@@ -45,6 +46,9 @@ interface BaseTransactionRepository<Entity : BaseTransaction> : BaseRepository<E
 
 @Repository
 interface TransferTransactionRepository : BaseTransactionRepository<TransferTransaction>
+
+@Repository
+interface CoinBaseTransactionRepository : BaseTransactionRepository<CoinBaseTransaction>
 
 @Repository
 interface VoteTransactionRepository : BaseTransactionRepository<VoteTransaction>