+++ src/main/kotlin/io/openfuture/chain/service/DefaultBlockService.kt
@@ -4,6 +4,7 @@ import io.openfuture.chain.entity.Block
 import io.openfuture.chain.entity.GenesisBlock
 import io.openfuture.chain.entity.MainBlock
 import io.openfuture.chain.entity.transaction.BaseTransaction
+import io.openfuture.chain.entity.transaction.CoinBaseTransaction
 import io.openfuture.chain.exception.NotFoundException
 import io.openfuture.chain.repository.BlockRepository
 import io.openfuture.chain.repository.GenesisBlockRepository
@@ -17,6 +18,7 @@ class DefaultBlockService(
     private val mainBlockRepository: MainBlockRepository,
     private val genesisBlockRepository: GenesisBlockRepository,
     private val transactionService: BaseTransactionService<BaseTransaction>,
+    private val coinBaseTransactionService: CoinBaseTransactionService,
     private val walletService: WalletService
 ) : BlockService {
 
@@ -42,6 +44,7 @@ class DefaultBlockService(
 
     @Transactional
     override fun save(block: MainBlock): MainBlock {
+        coinBaseTransactionService.save(block.transactions.first() as CoinBaseTransaction)
         val savedBlock = mainBlockRepository.save(block)
         val transactions = block.transactions
         for (transaction in transactions) {